refactor: 사진 촬영 페이지 코드 분리 및 정리
